Moved get_biolink_prefix_map() into node.py.
